refactor: extract shared waiver blocklist to src/lib (DRY fix)
Both pr-kaizen-clear.ts and reflection-checks.ts had identical
blocklists that would drift independently. Extract to single
source of truth in src/lib/waiver-blocklist.ts.
