+++ src/hooks/pr-kaizen-clear.ts
@@ -86,33 +86,9 @@ function logWaiver(
 
 // ── Waiver quality (kaizen #446) ─────────────────────────────────────
 
-/**
- * Generic no-action reasons that mask real friction.
- * Sourced from real incidents: #388, #280, #258.
- */
-const GENERIC_WAIVER_BLOCKLIST = [
-  'overengineering',
-  'over-engineering',
-  'low frequency',
-  'self-correcting',
-  'acceptable tradeoff',
-  'acceptable trade-off',
-  'not worth it',
-  'too complex',
-  'diminishing returns',
-  'edge case',
-  'minor issue',
-  'cosmetic',
-];
-
-/** Check if a no-action reason matches the blocklist. Returns matched term or null. */
-export function matchesWaiverBlocklist(reason: string): string | null {
-  const lower = reason.toLowerCase();
-  for (const blocked of GENERIC_WAIVER_BLOCKLIST) {
-    if (lower.includes(blocked)) return blocked;
-  }
-  return null;
-}
+// Re-export for test imports; single source of truth in src/lib/waiver-blocklist.ts
+export { matchesWaiverBlocklist } from '../lib/waiver-blocklist.js';
+import { matchesWaiverBlocklist } from '../lib/waiver-blocklist.js';
 
 // ── Validation ───────────────────────────────────────────────────────
